Actually on second thought, we can't keep the same interfaces across job gauge changes, so instead remove them entirely.
