+++ OverlayPlugin.Core/MemoryProcessors/JobGauge/JobGauge655.cs
@@ -69,302 +69,5 @@ namespace RainbowMage.OverlayPlugin.MemoryProcessors.JobGauge
                 return (JobGaugeManager)Marshal.PtrToStructure(new IntPtr(buffer), typeof(JobGaugeManager));
             }
         }
-
-        #region OverlayPlugin Interfaces
-
-        public interface IBaseJobGauge
-        {
-        }
-
-        #region Healer
-
-        public interface IBaseWhiteMageGauge : IBaseJobGauge
-        {
-            short LilyTimer { get; }
-            byte Lily { get; }
-            byte BloodLily { get; }
-        }
-
-        public interface IBaseScholarGauge : IBaseJobGauge
-        {
-            byte Aetherflow { get; }
-            byte FairyGauge { get; }
-            short SeraphTimer { get; }
-            byte DismissedFairy { get; }
-        }
-
-        public interface IBaseAstrologianGauge : IBaseJobGauge
-        {
-            short Timer { get; }
-            byte Card { get; }
-            byte Seals { get; }
-
-            AstrologianCard CurrentCard { get; }
-
-            AstrologianSeal[] CurrentSeals { get; }
-        }
-
-        public interface IBaseSageGauge : IBaseJobGauge
-        {
-            short AddersgallTimer { get; }
-            byte Addersgall { get; }
-            byte Addersting { get; }
-            byte Eukrasia { get; }
-
-            bool EukrasiaActive { get; }
-        }
-
-        #endregion
-
-        #region MagicDPS
-
-        public interface IBaseBlackMageGauge : IBaseJobGauge
-        {
-            short EnochianTimer { get; }
-            short ElementTimeRemaining { get; }
-            sbyte ElementStance { get; }
-            byte UmbralHearts { get; }
-            byte PolyglotStacks { get; }
-            EnochianFlags EnochianFlags { get; }
-
-            int UmbralStacks { get; }
-            int AstralStacks { get; }
-            bool EnochianActive { get; }
-            bool ParadoxActive { get; }
-        }
-
-        public interface IBaseSummonerGauge : IBaseJobGauge
-        {
-            ushort SummonTimer { get; }
-            ushort AttunementTimer { get; }
-            byte ReturnSummon { get; }
-            byte ReturnSummonGlam { get; }
-            byte Attunement { get; }
-            AetherFlags AetherFlags { get; }
-        }
-
-        public interface IBaseRedMageGauge : IBaseJobGauge
-        {
-            byte WhiteMana { get; }
-            byte BlackMana { get; }
-            byte ManaStacks { get; }
-        }
-
-        #endregion
-
-        #region RangeDPS
-
-        public interface IBaseBardGauge : IBaseJobGauge
-        {
-            ushort SongTimer { get; }
-            byte Repertoire { get; }
-            byte SoulVoice { get; }
-            SongFlags SongFlags { get; }
-        }
-
-        public interface IBaseMachinistGauge : IBaseJobGauge
-        {
-            short OverheatTimeRemaining { get; }
-            short SummonTimeRemaining { get; }
-            byte Heat { get; }
-            byte Battery { get; }
-            byte LastSummonBatteryPower { get; }
-            byte TimerActive { get; }
-        }
-
-        public interface IBaseDancerGauge : IBaseJobGauge
-        {
-            byte Feathers { get; }
-            byte Esprit { get; }
-            byte[] DanceSteps { get; }
-            byte StepIndex { get; }
-
-            DanceStep CurrentStep { get; }
-        }
-
-        #endregion
-
-        #region MeleeDPS
-
-        public interface IBaseMonkGauge : IBaseJobGauge
-        {
-            byte Chakra { get; }
-
-            BeastChakraType BeastChakra1 { get; }
-
-            BeastChakraType BeastChakra2 { get; }
-
-            BeastChakraType BeastChakra3 { get; }
-
-            NadiFlags Nadi { get; }
-            ushort BlitzTimeRemaining { get; }
-
-            BeastChakraType[] BeastChakra { get; }
-        }
-
-        public interface IBaseDragoonGauge : IBaseJobGauge
-        {
-            short LotdTimer { get; }
-            byte LotdState { get; }
-            byte EyeCount { get; }
-            byte FirstmindsFocusCount { get; }
-        }
-
-        public interface IBaseNinjaGauge : IBaseJobGauge
-        {
-            ushort HutonTimer { get; }
-            byte Ninki { get; }
-            byte HutonManualCasts { get; }
-        }
-
-        public interface IBaseSamuraiGauge : IBaseJobGauge
-        {
-            KaeshiAction Kaeshi { get; }
-            byte Kenki { get; }
-            byte MeditationStacks { get; }
-            SenFlags SenFlags { get; }
-        }
-
-        public interface IBaseReaperGauge : IBaseJobGauge
-        {
-            byte Soul { get; }
-            byte Shroud { get; }
-            ushort EnshroudedTimeRemaining { get; }
-            byte LemureShroud { get; }
-            byte VoidShroud { get; }
-        }
-
-        #endregion
-
-        #region Tanks
-
-        public interface IBaseDarkKnightGauge : IBaseJobGauge
-        {
-            byte Blood { get; }
-            ushort DarksideTimer { get; }
-            byte DarkArtsState { get; }
-            ushort ShadowTimer { get; }
-        }
-
-        public interface IBasePaladinGauge : IBaseJobGauge
-        {
-            byte OathGauge { get; }
-        }
-
-        public interface IBaseWarriorGauge : IBaseJobGauge
-        {
-            byte BeastGauge { get; }
-        }
-
-        public interface IBaseGunbreakerGauge : IBaseJobGauge
-        {
-            byte Ammo { get; }
-            short MaxTimerDuration { get; }
-            byte AmmoComboStep { get; }
-        }
-
-        public enum AstrologianCard
-        {
-            None = 0,
-            Balance = 1,
-            Bole = 2,
-            Arrow = 3,
-            Spear = 4,
-            Ewer = 5,
-            Spire = 6,
-            Lord = 0x70,
-            Lady = 0x80
-        }
-
-        public enum AstrologianSeal
-        {
-            Solar = 1,
-            Lunar = 2,
-            Celestial = 3
-        }
-
-        public enum DanceStep : byte
-        {
-            Finish = 0,
-            Emboite = 1,
-            Entrechat = 2,
-            Jete = 3,
-            Pirouette = 4
-        }
-
-        [Flags]
-        public enum EnochianFlags : byte
-        {
-            None = 0,
-            Enochian = 1,
-            Paradox = 2
-        }
-
-        public enum KaeshiAction : byte
-        {
-            Higanbana = 1,
-            Goken = 2,
-            Setsugekka = 3,
-            Namikiri = 4
-        }
-
-        [Flags]
-        public enum SenFlags : byte
-        {
-            None = 0,
-            Setsu = 1 << 0,
-            Getsu = 1 << 1,
-            Ka = 1 << 2
-        }
-
-        [Flags]
-        public enum SongFlags : byte
-        {
-            None = 0,
-            MagesBallad = 1 << 0,
-            ArmysPaeon = 1 << 1,
-            WanderersMinuet = MagesBallad | ArmysPaeon,
-            MagesBalladLastPlayed = 1 << 2,
-            ArmysPaeonLastPlayed = 1 << 3,
-            WanderersMinuetLastPlayed = MagesBalladLastPlayed | ArmysPaeonLastPlayed,
-            MagesBalladCoda = 1 << 4,
-            ArmysPaeonCoda = 1 << 5,
-            WanderersMinuetCoda = 1 << 6
-        }
-
-        [Flags]
-        public enum AetherFlags : byte
-        {
-            None = 0,
-            Aetherflow1 = 1 << 0,
-            Aetherflow2 = 1 << 1,
-            Aetherflow = Aetherflow1 | Aetherflow2,
-            IfritAttuned = 1 << 2,
-            TitanAttuned = 1 << 3,
-            GarudaAttuned = TitanAttuned | IfritAttuned,
-            PhoenixReady = 1 << 4,
-            IfritReady = 1 << 5,
-            TitanReady = 1 << 6,
-            GarudaReady = 1 << 7
-        }
-
-        public enum BeastChakraType : byte
-        {
-            None = 0,
-            Coeurl = 1,
-            OpoOpo = 2,
-            Raptor = 3
-        }
-
-        [Flags]
-        public enum NadiFlags : byte
-        {
-            Lunar = 2,
-            Solar = 4
-        }
-
-        #endregion
-        #endregion
-
     }
 }
